book_writer/requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
pydantic==2.6.3
langchain==0.1.12
langchain-community==0.0.28
google-generativeai==0.4.1
langchain-google-genai==0.0.11 